setup.py: use XCode include too on macOS
macOS does not really play with the Unix structure and many of the
headers is only found under XCode in newer versions. This commit adds
a search there.

The whole "grep the files" thing is obviously inferior to using stuff
that spawns compilers for testing (like cmake and autoconf), but I don't
even think such a thing exists in the Python world.